draft_combine endpoint and tests
